// public/enums/TruckStatus.ts
export enum TruckStatus {
    AVAILABLE = 'AVAILABLE',
    UNDER_MAINTENANCE = 'UNDER_MAINTENANCE',
    OUT_OF_SERVICE = 'OUT_OF_SERVICE',
    IN_TRANSIT = 'IN_TRANSIT',
    LOADING = 'LOADING',
    UNLOADING = 'UNLOADING',
    WAITING_FOR_ASSIGNMENT = 'WAITING_FOR_ASSIGNMENT',
    IDLE = 'IDLE',
    ON_SALE = 'ON_SALE'
  }